+++ CHANGELOG.md
@@ -30,8 +30,7 @@ You can check your current version with the following command:
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
 
-### 2.4.0a1
-**June 25, 2025**
+### June 25, 2025 [minor]
 - Added `FeasibilityStudy` active record class to `tasking` module.
 
 ### June 24, 2025 [patch]